requirements.txt
ultralytics
opencv-python
Pillow
numpy
tqdm
scikit-image
streamlit